+++ packages/eyes-sdk-core/test/testUtils.js
@@ -4,6 +4,7 @@
 const fs = require('fs')
 const path = require('path')
 const {EyesBase} = require('../index')
+const makeCheck = require('../lib/new/check')
 
 const getResourcePath = fileName => path.resolve(__dirname, 'fixtures', fileName)
 
@@ -92,6 +93,45 @@ function resetEnvVars() {
   })
 }
 
+function makeSetCut({ eyes }) {
+  return async function setCut(cut) {
+    eyes.setCut(cut)
+  }
+}
+function makeGetCut({ eyes }) {
+  return async function getCut() {
+    return eyes.getCut()
+  }
+}
+class EyesMock {
+  constructor() {
+    this.appName = 'App',
+      this.testName = 'TestImageCut'
+    this._configuration = {
+      mergeConfig: () => { }
+    }
+  }
+  check() {
+    return this;
+  }
+  setCut(cut) {
+    this.cut = cut;
+  }
+  getCut() {
+    return this.cut;
+  }
+  toJSON() { }
+}
+function EyesMockWithCutFunc() {
+  const eyes = new EyesMock()
+  return {
+    check: makeCheck({ eyes }),
+    setCut: makeSetCut({ eyes }),
+    getCut: makeGetCut({ eyes }),
+  }
+}
+
+
 exports.getResourcePath = getResourcePath
 exports.getResource = getResource
 exports.getResourceAsText = getResourceAsText
@@ -101,3 +141,4 @@ exports.imageMock2String = imageMock2String
 exports.getResourcePath = getResourcePath
 exports.getResource = getResource
 exports.resetEnvVars = resetEnvVars
+exports.EyesMockWithCutFunc = EyesMockWithCutFunc