+++ packages/screenshoter/test/e2e/android/app-fully-scroll.spec.js
@@ -1,0 +1,33 @@
+const {makeDriver, sleep, test} = require('../e2e')
+
+describe('screenshoter android app', () => {
+  const logger = {log: () => {}, warn: () => {}, error: () => {}, verbose: () => {}}
+  let driver, destroyDriver
+
+  before(async () => {
+    ;[driver, destroyDriver] = await makeDriver({type: 'android', logger})
+  })
+
+  after(async () => {
+    await destroyDriver()
+  })
+
+  it('take full app screenshot on screen with scroll view', async () => {
+    const button = await driver.element({type: 'id', selector: 'btn_scroll_view_footer_header'})
+    await button.click()
+    await sleep(3000)
+
+    await driver.init()
+
+    await test({
+      type: 'android',
+      tag: 'app-fully-scroll',
+      fully: true,
+      framed: true,
+      wait: 1500,
+      scrollingMode: 'scroll',
+      driver,
+      logger,
+    })
+  })
+})