+++ packages/screenshoter/test/e2e/ios/app-statusbar.spec.js
@@ -1,0 +1,31 @@
+const {makeDriver, sleep, test} = require('../e2e')
+
+describe('screenshoter ios app', () => {
+  const logger = {log: () => {}, warn: () => {}, error: () => {}, verbose: () => {}}
+  let driver, destroyDriver
+
+  before(async () => {
+    ;[driver, destroyDriver] = await makeDriver({type: 'ios', logger})
+  })
+
+  after(async () => {
+    await destroyDriver()
+  })
+
+  it('take viewport screenshot with status bar', async () => {
+    const button = await driver.element({type: 'accessibility id', selector: 'Empty table view'})
+    await button.click()
+    await sleep(3000)
+
+    await driver.init()
+
+    await test({
+      type: 'ios',
+      tag: 'app',
+      withStatusBar: true,
+      wait: 1500,
+      driver,
+      logger,
+    })
+  })
+})