+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## Unreleased
 - updated to @applitools/types@1.0.24 (from 1.0.23)
-
+- allow setting 'setImageCut' also after eyes.open
 
 
 ## 13.0.0 - 2022/1/19