+++ packages/eyes-sdk-core/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 - updated to @applitools/screenshoter@3.3.5 (from 3.3.4)
+- updated to @applitools/driver@1.4.11 (from 1.4.10)
 
 
 ## 13.0.1 - 2022/2/4